+++ modules/material-management-browser/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.1.0](https://github.com/aws/aws-encryption-sdk-javascript/compare/v3.0.3...v3.1.0) (2021-11-10)
+
+**Note:** Version bump only for package @aws-crypto/material-management-browser
+
+
+
+
+
+## [3.0.2](https://github.com/aws/aws-encryption-sdk-javascript/compare/v3.0.0...v3.0.2) (2021-09-08)
+
+**Note:** Version bump only for package @aws-crypto/material-management-browser
+
+
+
+
+
+## [3.0.1](https://github.com/aws/aws-encryption-sdk-javascript/compare/v3.0.0...v3.0.1) (2021-09-07)
+
+**Note:** Version bump only for package @aws-crypto/material-management-browser
+
+
+
+
+
 # [3.0.0](https://github.com/aws/aws-encryption-sdk-javascript/compare/v2.4.0...v3.0.0) (2021-07-14)
 
 **Note:** Version bump only for package @aws-crypto/material-management-browser